Day 4: Class vs. Instance
Method
A sort of named procedure associated with a class that performs a predefined action. In the sample code below, returnType will either be a data type or  if no value need be returned. Like a constructor, a method can have  or more parameters.

returnType methodName(parameterOne, ..., parameterN){
    ...
    return variableOfReturnType; // no return statement if void
}
Most classes will have methods called getters and setters that get (return) or set the values of its instance variables. Standard getter/setter syntax:

class MyClass{
    dataType instanceVariable;
    ...
    void setInstanceVariable(int value){
        this.instanceVariable = value;
    }
    dataType getInstanceVariable(){
        return instanceVariable;
    }
}
Structuring code this way is a means of managing how the instance variable is accessed and/or modified.

Parameter
A parenthetical variable in a function or constructor declaration (e.g.: in int methodOne(int x), the parameter is int x).

Argument
The actual value of a parameter (e.g.: in methodOne(5), the argument passed as variable x is 5).