Day 7 of Code: Arrays!
Data Structures
A way of organizing data that enables efficient storage, retrieval, and use.

Arrays
A type of data structure that stores elements of the same type (generally). It's important to note that you'll often see arrays referred to as  in documentation, but the variable names you use when coding should be descriptive and begin with lowercase letters.