Day 8: Dictionaries and Maps
Java Maps
Map is an interface that provides a blueprint for data structures that take  pairs and map keys to their associated values (it's important to note that both the  and the  must be Objects and not primitives). The implementation is done by implementing classes such as HashMap or LinkedHashMap. Consider the following code:

// Declare a String to String map
Map<String, String> myMap; 

// Initialize it as a new String to String HashMap
myMap = new HashMap<String, String>();  

// Change myMap to be a new (completely different) String to String LinkedHashMap instead
myMap = new LinkedHashMap<String, String>();